[ERSSUP-78354]-[]-[Moved get_env to new utils.py to avoid cyclic dependency]-[JW]
