+++ tests/utils.py
@@ -1,0 +1,12 @@
+import os
+
+
+def get_env(variable_name: str) -> str:
+    """Returns an environment variable"""
+    try:
+        var = os.environ[variable_name]
+        if not var:
+            raise RuntimeError(f"Variable is null, Check {variable_name}.")
+        return var
+    except KeyError:
+        raise RuntimeError(f"Variable is not set, Check {variable_name}.")
